+++ src/systems/dnd5e.js
@@ -1,5 +1,6 @@
 import GiveItems from "../applications/dialogs/give-items-dialog/give-items-dialog.js";
 import PrivateAPI from "../API/private-api.js";
+import CONSTANTS from "../constants/constants.js";
 
 export default {
 
@@ -12,7 +13,7 @@ export default {
 	"ITEM_CLASS_LOOT_TYPE": "loot",
 
 	// The item class type is the type of item that will be used for the default weapon item
-	"ITEM_CLASS_WEAPON_TYPE": "weapon", 
+	"ITEM_CLASS_WEAPON_TYPE": "weapon",
 
 	// The item class type is the type of item that will be used for the default equipment item
 	"ITEM_CLASS_EQUIPMENT_TYPE": "equipment",
@@ -35,6 +36,8 @@ export default {
 		}
 	],
 
+	"UNSTACKABLE_ITEM_TYPES": ["container"],
+
 	// This function is an optional system handler that specifically transforms an item when it is added to actors
 	"ITEM_TRANSFORMER": async (itemData) => {
 		["equipped", "proficient", "prepared"].forEach(key => {
@@ -51,6 +54,25 @@ export default {
 			}
 		}
 		return itemData;
+	},
+
+	"ITEM_TYPE_HANDLERS": {
+		"GLOBAL": {
+			[CONSTANTS.ITEM_TYPE_METHODS.IS_CONTAINED]: ({ item }) => {
+				return item.system.container;
+			}
+		},
+		"container": {
+			[CONSTANTS.ITEM_TYPE_METHODS.HAS_CURRENCY]: true,
+			[CONSTANTS.ITEM_TYPE_METHODS.CONTENTS]: ({ item }) => {
+				return item.system.contents.contents;
+			},
+			[CONSTANTS.ITEM_TYPE_METHODS.TRANSFER]: ({ item, items }) => {
+				for (const containedItem of item.system.contents.contents) {
+					items.push(containedItem.toObject());
+				}
+			}
+		}
 	},
 
 	// This function is an optional system handler that specifically transforms an item's price into a more unified numeric format